+++ ADFastAPI/main.py
@@ -3,7 +3,7 @@ import uvicorn
 import httpx
 
 from sources.api.make_ad import step1, step2, step3, step4
-from sources.api import add_animation
+from sources.api import configure_animation
 
 from starlette.exceptions import HTTPException
 from fastapi.exceptions import RequestValidationError
@@ -34,7 +34,7 @@ app.include_router(step1.router)
 app.include_router(step2.router)
 app.include_router(step3.router)
 app.include_router(step4.router)
-app.include_router(add_animation.router)
+app.include_router(configure_animation.router)
 
 if __name__ == '__main__':
     uvicorn.run('main:app', host='0.0.0.0', port=8000, reload=True)
